+++ CHANGELOG.md
@@ -12,7 +12,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Fixed** for any bug fixes.
 - **Security** in case of vulnerabilities.
 
-## [Unreleased] - Started: 2-11-20
+## [Unreleased] - Started: 2-25-20
+
+## [1.2.3] - Started: 2-11-20
+
+### Added
+- `Orion`: Build out custom hook `useAquila` to make interfacing w/ `Aquila-js` api easier.
+- `Orion`: Built out custom hook `useBom` (business object modal).
+
+### Changed
+- `Phoenix`: Show team and client names for Update Role Sitemap operation in OpLog
+- `Polaris`: Upgrade `react-router-dom` and `react-router` to gain access to useLocation and useHistory hooks.
+
+### Fixed
+- `Phoenix`: Trim empty spaces on user's username and email on create/update
+- `Phoenix`: Prevent plain text passwords from being logged to api.log file
 
 ## [1.2.2] - Started: 2-11-20